Added logic to handle exception

+++ app/src/main/java/a23sokolov/bankcharts/view/MainActivity.kt
@@ -21,7 +21,7 @@ class MainActivity : BaseActivity<MainViewModel>() {
     private fun initViewModel(viewModel: MainViewModel) {
         afterViewModelInited(viewModel)
         downloadBtn.setOnClickListener { view ->
-            viewModel.getPointsBtnClicked(countPoints.text.toString().toInt())
+            viewModel.getPointsBtnClicked(countPoints.text.toString())
         }
     }
